Refactor types to be more truthful
