+++ server/src/controllers/conversationController.ts
@@ -52,10 +52,10 @@ const createConversation = asyncHandler(async (req: Request, res: Response) => {
 
     let newConversation;
     if (type == 1) {
-      newConversation = new Conversation({ type: "1" });
+      newConversation = new Conversation({ type: 1 });
     } else {
       newConversation = new Conversation({
-        type: "2",
+        type: 2,
         owner: id,
         name: "New Group Chat",
       });
@@ -95,7 +95,7 @@ const postImage = asyncHandler(
     const { file } = req;
 
     const conversations = await ConversationModel.findExistingConversations(
-      userId
+      Number(userId)
     );
 
     if (conversations === null) {
@@ -109,7 +109,7 @@ const postImage = asyncHandler(
 
       const message = new Message({
         author: userId,
-        conversation: conversationId,
+        conversation: Number(conversationId),
         media_url: media_url as string,
         description: "Image by " + username,
         type: 3,
@@ -153,7 +153,7 @@ const getConversation = asyncHandler(async (req: Request, res: Response) => {
   const { id: userId } = req.session.user;
 
   const conversation = await ConversationModel.findExistingConversation(
-    conversationId
+    Number(conversationId)
   );
 
   if (!conversation) {
